backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
argon2-cffi
